requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.6.0
timm==0.9.12
Pillow==10.1.0
gdown==4.7.1
requests==2.31.0
numpy==1.26.1
aiohttp==3.9.1
diskcache==5.6.3
tqdm==4.66.1
